yolo_first_round_opt/requirements.txt
ultralytics>=8.2.0
sahi>=0.11.18
pyyaml>=6.0
